% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge_observations_aspe.R
\name{charge_observations_aspe}
\alias{charge_observations_aspe}
\title{charge_observations_aspe}
\usage{
charge_observations_aspe(code_operation)
}
\arguments{
\item{code_operation}{Vecteur avec les codes opérations à charger (paramètre obligatoire)}
}
\value{
Liste des stations ciblées
}
\description{
charge_observations_aspe
}
\examples{

obs<-charge_observations_aspe(code_operation=c("89771","89805"))

}
