% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge_stations_aspe.R
\name{charge_stations_aspe}
\alias{charge_stations_aspe}
\title{charge_stations_aspe}
\usage{
charge_stations_aspe(dpt, code_agence = c("04"))
}
\arguments{
\item{dpt}{Vecteur avec les départements à charger}

\item{code_agence}{Vecteur avec les codes agences selon https://id.eaufrance.fr/nsa/447}
}
\value{
Liste des stations ciblées
}
\description{
charge_stations_aspe
}
\examples{

dpt<-c("44", "56", "22", "35", "49", "53")
liste_stations<-charge_stations_aspe(dpt, code_agence = c("04"))


}
