% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liste_operations_aspe.R
\name{liste_operations_aspe}
\alias{liste_operations_aspe}
\title{liste_operations_aspe}
\usage{
liste_operations_aspe(
  codes_stations = NULL,
  code_point_prelevement_aspe = NULL,
  date_debut = NULL,
  date_fin = NULL,
  dpt = NULL,
  code_agence = NULL
)
}
\arguments{
\item{codes_stations}{Vecteur avec les codes stations dont on veut les opérations.}

\item{code_point_prelevement_aspe}{Vecteur avec les codes stations dont on veut les opérations}

\item{date_debut}{Date minimale de création de l'opération en base de données. (format date)}

\item{date_fin}{Date maximale de création de l'opération en base de données (format date)}

\item{dpt}{Vecteur avec les départements à charger}

\item{code_agence}{Vecteur avec les codes agences selon https://id.eaufrance.fr/nsa/447}
}
\value{
Liste des operations
}
\description{
liste_operations_aspe
}
\examples{

dpt<-c("44", "56")
liste_stations<-charge_stations_aspe(dpt, code_agence = c("04"))
liste_stations<-liste_stations[!is.na(liste_stations$code_station),]
liste_op<-liste_operations_aspe(codes_stations=head(liste_stations$code_station,10), 
                      date_debut=as.Date("2010-01-01"), 
                      date_fin=as.Date("2024-01-01")) 

liste_op2<-liste_operations_aspe(code_point_prelevement_aspe = head(liste_stations$code_point_prelevement_aspe,10),                       date_debut=as.Date("2010-01-01"), 
                      date_fin=as.Date("2024-01-01")) 


}
