% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_observations_aspe.R
\name{split_observations_aspe}
\alias{split_observations_aspe}
\title{split_observations_aspe}
\usage{
split_observations_aspe(observations)
}
\arguments{
\item{observations}{Résultat renvoyé par la fonction charge_observations_aspe}
}
\value{
Liste avec d'un côté les conditions de l'opération et de l'autre les listes faunistiques
}
\description{
split_observations_aspe
}
\examples{

obs<-charge_observations_aspe(code_operation=c("89771","89805"))
detail<-split_observations_aspe(obs)


}
